Remove `--partial` flag from config imports.
The `--partial` flag in config imports prevents the import from removing
existing configuration that is no longer represented in the exported
configuration, but this could easily result in drift from expected
behavior when configuration is changed or revised.

This command also fails during deployments if the `config` module is not
explicitly enabled adding another dependency that isn't commonly enabled
in existing builds.