Refactor inheritance of sub-generator stage methods.
Implement these methods as abstract to enforce visibility of them on
sub-generators when it comes time for their methods to be queued as
tasks by Yeoman for execution.

By default, Yeoman is not seeing inherited methods since they are
captured on the object prototype instead of the instance directly.

Implementation of the inherited functionality for these methods has
been provided in prefixed `_do<Stage>` methods to simplify the
generator-level implementation of these methods to little more than
a public declaration of the method to make it discoverable.