The web directory is not created or needs to be removed.
Since the create-project command no longer completes the whole
installation and scaffolding process, the assumed `web` directory for
the web root built into the composer projects is not auto-created.
Executing the renaming operation before the next install operation that
finishes the scaffolding process allows specification of what the web
root should be renamed to, so the final `web` directory is no longer
created and no longer needs to be removed.